requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
arabic-reshaper==3.0.0
argcomplete==3.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
boto3==1.37.15
botocore==1.37.15
Brotli==1.1.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
cryptography==42.0.8
cssselect2==0.7.0
cssutils==2.11.1
dj-database-url==2.2.0
Django==5.0.4
django-autoslug==1.9.9
django-celery-beat==2.7.0
django-crispy-forms==2.1
django-dash==0.6.1
django-inlinecss==0.3.0
django-minio-storage==0.5.7
django-nine==0.2.7
django-redis==5.4.0
django-redis-cache==3.0.1
django-storages==1.14.2
django-timezone-field==7.0
django-tinymce==4.0.0
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
easy-thumbnails==2.8.5
et_xmlfile==2.0.0
feedparser==6.0.11
fonttools==4.53.0
frozenlist==1.4.1
future==1.0.0
gunicorn==22.0.0
html5lib==1.1
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.4.1
lxml==5.2.2
Markdown==3.6
MarkupSafe==2.1.5
minio==7.2.6
more-itertools==10.3.0
multidict==6.0.5
mysqlclient==2.2.4
num2words==0.5.13
numpy==2.1.1
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.0
pandas==2.2.3
pif==0.8.2
pillow==10.4.0
pipx==1.5.0
platformdirs==4.2.1
prompt_toolkit==3.0.47
psycopg==3.1.18
psycopg2==2.9.9
pycparser==2.22
pycryptodome==3.20.0
pydyf==0.10.0
Pygments==2.18.0
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pynliner==0.8.0
pypdf==4.2.0
pyphen==0.15.0
pypng==0.20220715.0
python-bidi==0.4.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
redis==3.5.3
reportlab==4.2.0
requests==2.31.0
s3transfer==0.11.4
setuptools==76.0.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
transliterate==1.10.2
twilio==9.0.5
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
Unidecode==1.3.8
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.1
userpath==1.9.2
vine==5.1.0
vishap==0.1.5
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
wheel==0.45.1
whitenoise==6.7.0
xhtml2pdf==0.2.16
XlsxWriter==3.2.1
yarl==1.9.4
zopfli==0.2.3